removed commented codes, added confirmation page, replaced button widgets with elevated button
