--- lib/providers/theme_provider.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ThemeProvider extends ChangeNotifier {
-  ThemeMode themeMode = ThemeMode.dark;
-
-  bool get isDarkMode => themeMode == ThemeMode.dark;
-
-  void toggleTheme(bool isOn) {
-    themeMode = isOn ? ThemeMode.dark : ThemeMode.light;
-    notifyListeners();
-  }
-}
-
-class MyThemes {
-  static final darkTheme = ThemeData(
-      scaffoldBackgroundColor: Colors.grey[900],
-      colorScheme: ColorScheme.dark());
-
-  static final lightTheme = ThemeData(
-      scaffoldBackgroundColor: Colors.grey[300],
-      colorScheme: ColorScheme.light());
-}
